requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

requests==2.26.0
azure-functions==1.7.1
azure-storage-blob==12.9.0
polars==0.20.10